backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.1.4
google-generativeai==0.8.3
python-multipart==0.0.6
httpx==0.26.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2024.1

# Payment dependencies
stripe==8.0.0

